// src/types/export.models.ts
export enum ExportModel {
  Users = "Users",
  Providers = "Providers",
  Listings = "Listings",
  Categories = "Categories",
  Cities = "Cities",
  Countries = "Countries",
  Banks = "Banks",
  DeliveryCompanies = "Delivery Companies",
  Contracts = "Contracts",
  Payments = "Payments",
  Ratings = "Ratings",
  Complaints = "Complaints",
  Notifications = "Notifications",
  FAQs = "FAQs",
  ContactMessages = "Contact Messages",
  SignedContracts = "Signed Contracts",
  Favorites = "Favorites",
  Conversations = "Conversations",
  Tracking = "Tracking",
  Admins = "Admins",
  Permissions = "Permissions",
  AdminPermissions = "Admin Permissions",
  Settings = "Settings",
}